refactor: remove old download preferences
